microservices/obtener_usuario/requirements.txt
pyodbc==5.1.0
Flask==3.0.3
pyjwt==2.8.0